+++ test/e2e/data/e2e_conf.yaml
@@ -17,11 +17,19 @@ images:
   # Use local dev images built source tree;
   - name: gcr.io/k8s-staging-cluster-api/capa-manager:e2e
     loadBehavior: mustLoad
-  - name: quay.io/jetstack/cert-manager-cainjector:v0.16.1
+
+## PLEASE KEEP THESE UP TO DATE WITH THE COMPONENTS
+  - name: quay.io/jetstack/cert-manager-cainjector:v1.1.0
     loadBehavior: tryLoad
-  - name: quay.io/jetstack/cert-manager-webhook:v0.16.1
+  - name: quay.io/jetstack/cert-manager-webhook:v1.1.0
     loadBehavior: tryLoad
-  - name: quay.io/jetstack/cert-manager-controller:v0.16.1
+  - name: quay.io/jetstack/cert-manager-controller:v1.1.0
+    loadBehavior: tryLoad
+  - name: k8s.gcr.io/cluster-api/kubeadm-control-plane-controller:v0.4.0
+    loadBehavior: tryLoad
+  - name: k8s.gcr.io/cluster-api/kubeadm-bootstrap-controller:v0.4.0
+    loadBehavior: tryLoad
+  - name: k8s.gcr.io/cluster-api/cluster-api-controller:v0.4.0
     loadBehavior: tryLoad
 
 providers:
@@ -152,7 +160,7 @@ intervals:
   conformance/wait-worker-nodes: ["30m", "10s"]
   default/wait-controllers: ["3m", "10s"]
   default/wait-delete-cluster: ["20m", "10s"]
-  default/wait-machine-upgrade: ["20m", "10s"]
+  default/wait-machine-upgrade: ["30m", "10s"]
   default/wait-machine-status: ["20m", "10s"]
   default/wait-failed-machine-status: ["2m", "10s"]
   default/wait-infra-subnets: ["5m", "30s"]
