+++ test/e2e/suites/conformance/conformance_test.go
@@ -24,7 +24,7 @@ import (
 	"path/filepath"
 	"strconv"
 
-	. "github.com/onsi/ginkgo"
+	"github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/utils/pointer"
@@ -35,7 +35,8 @@ import (
 	"sigs.k8s.io/cluster-api/util"
 )
 
-var _ = Describe("conformance tests", func() {
+// TODO @randomvariable: Replace with CAPI e2e framework ClusterUpgradeConformanceSpec
+var _ = ginkgo.Describe("[unmanaged] [conformance] tests", func() {
 	var (
 		namespace *corev1.Namespace
 		ctx       context.Context
@@ -43,7 +44,7 @@ var _ = Describe("conformance tests", func() {
 		result    *clusterctl.ApplyClusterTemplateAndWaitResult
 	)
 
-	BeforeEach(func() {
+	ginkgo.BeforeEach(func() {
 		Expect(e2eCtx.Environment.BootstrapClusterProxy).ToNot(BeNil(), "Invalid argument. BootstrapClusterProxy can't be nil")
 		Expect(e2eCtx.E2EConfig).ToNot(BeNil(), "Invalid argument. e2eConfig can't be nil when calling %s spec", specName)
 		Expect(e2eCtx.E2EConfig.Variables).To(HaveKey(shared.KubernetesVersion))
@@ -52,7 +53,7 @@ var _ = Describe("conformance tests", func() {
 		namespace = shared.SetupSpecNamespace(ctx, specName, e2eCtx)
 		result = new(clusterctl.ApplyClusterTemplateAndWaitResult)
 	})
-	Measure(specName, func(b Benchmarker) {
+	ginkgo.Measure(specName, func(b ginkgo.Benchmarker) {
 
 		name := fmt.Sprintf("cluster-%s", util.RandomString(6))
 		shared.SetEnvVar("USE_CI_ARTIFACTS", "true", false)
@@ -103,7 +104,7 @@ var _ = Describe("conformance tests", func() {
 		b.RecordValue("conformance suite run time", runtime.Seconds())
 	}, 1)
 
-	AfterEach(func() {
+	ginkgo.AfterEach(func() {
 		shared.SetEnvVar("USE_CI_ARTIFACTS", "false", false)
 		// Dumps all the resources in the spec namespace, then cleanups the cluster object and the spec namespace itself.
 		shared.DumpSpecResourcesAndCleanup(ctx, "", namespace, e2eCtx)
