+++ test/e2e/suites/managed/eks_test.go
@@ -22,7 +22,7 @@ import (
 	"context"
 	"fmt"
 
-	. "github.com/onsi/ginkgo"
+	"github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	corev1 "k8s.io/api/core/v1"
@@ -34,7 +34,7 @@ import (
 )
 
 // General EKS e2e test
-var _ = Describe("EKS cluster tests", func() {
+var _ = ginkgo.Describe("[managed] [general] EKS cluster tests", func() {
 	var (
 		namespace       *corev1.Namespace
 		ctx             context.Context
@@ -45,7 +45,7 @@ var _ = Describe("EKS cluster tests", func() {
 	)
 
 	shared.ConditionalIt(runGeneralTests, "should create a cluster and add nodes", func() {
-		By("should have a valid test configuration")
+		ginkgo.By("should have a valid test configuration")
 		Expect(e2eCtx.Environment.BootstrapClusterProxy).ToNot(BeNil(), "Invalid argument. BootstrapClusterProxy can't be nil")
 		Expect(e2eCtx.E2EConfig).ToNot(BeNil(), "Invalid argument. e2eConfig can't be nil when calling %s spec", specName)
 		Expect(e2eCtx.E2EConfig.Variables).To(HaveKey(shared.KubernetesVersion))
@@ -54,19 +54,19 @@ var _ = Describe("EKS cluster tests", func() {
 		namespace = shared.SetupSpecNamespace(ctx, specName, e2eCtx)
 		clusterName = fmt.Sprintf("cluster-%s", util.RandomString(6))
 
-		By("setting up AWS statis credentials")
+		ginkgo.By("setting up AWS static credentials")
 		shared.SetupStaticCredentials(ctx, namespace, e2eCtx)
 
-		By("default iam role should exist")
-		verifyRoleExistsAndOwned(controlplanev1.DefaultEKSControlPlaneRole, clusterName, false, e2eCtx.BootstratpUserAWSSession)
+		ginkgo.By("default iam role should exist")
+		verifyRoleExistsAndOwned(controlplanev1.DefaultEKSControlPlaneRole, clusterName, false, e2eCtx.BootstrapUserAWSSession)
 
-		By("should create an EKS control plane")
+		ginkgo.By("should create an EKS control plane")
 		ManagedClusterSpec(ctx, func() ManagedClusterSpecInput {
 			return ManagedClusterSpecInput{
 				E2EConfig:                e2eCtx.E2EConfig,
 				ConfigClusterFn:          defaultConfigCluster,
 				BootstrapClusterProxy:    e2eCtx.Environment.BootstrapClusterProxy,
-				AWSSession:               e2eCtx.BootstratpUserAWSSession,
+				AWSSession:               e2eCtx.BootstrapUserAWSSession,
 				Namespace:                namespace,
 				ClusterName:              clusterName,
 				Flavour:                  EKSControlPlaneOnlyWithAddonFlavor,
@@ -76,12 +76,12 @@ var _ = Describe("EKS cluster tests", func() {
 			}
 		})
 
-		By("should have the VPC CNI installed")
+		ginkgo.By("should have the VPC CNI installed")
 		CheckAddonExistsSpec(ctx, func() CheckAddonExistsSpecInput {
 			return CheckAddonExistsSpecInput{
 				E2EConfig:             e2eCtx.E2EConfig,
 				BootstrapClusterProxy: e2eCtx.Environment.BootstrapClusterProxy,
-				AWSSession:            e2eCtx.BootstratpUserAWSSession,
+				AWSSession:            e2eCtx.BootstrapUserAWSSession,
 				Namespace:             namespace,
 				ClusterName:           clusterName,
 				AddonName:             cniAddonName,
@@ -89,13 +89,13 @@ var _ = Describe("EKS cluster tests", func() {
 			}
 		})
 
-		By("should create a MachineDeployment")
+		ginkgo.By("should create a MachineDeployment")
 		MachineDeploymentSpec(ctx, func() MachineDeploymentSpecInput {
 			return MachineDeploymentSpecInput{
 				E2EConfig:             e2eCtx.E2EConfig,
 				ConfigClusterFn:       defaultConfigCluster,
 				BootstrapClusterProxy: e2eCtx.Environment.BootstrapClusterProxy,
-				AWSSession:            e2eCtx.BootstratpUserAWSSession,
+				AWSSession:            e2eCtx.BootstrapUserAWSSession,
 				Namespace:             namespace,
 				ClusterName:           clusterName,
 				Replicas:              1,
@@ -103,13 +103,13 @@ var _ = Describe("EKS cluster tests", func() {
 			}
 		})
 
-		By("should create a managed node pool and scale")
+		ginkgo.By("should create a managed node pool and scale")
 		ManagedMachinePoolSpec(ctx, func() ManagedMachinePoolSpecInput {
 			return ManagedMachinePoolSpecInput{
 				E2EConfig:             e2eCtx.E2EConfig,
 				ConfigClusterFn:       defaultConfigCluster,
 				BootstrapClusterProxy: e2eCtx.Environment.BootstrapClusterProxy,
-				AWSSession:            e2eCtx.BootstratpUserAWSSession,
+				AWSSession:            e2eCtx.BootstrapUserAWSSession,
 				Namespace:             namespace,
 				ClusterName:           clusterName,
 				IncludeScaling:        true,
@@ -134,7 +134,7 @@ var _ = Describe("EKS cluster tests", func() {
 			Cluster: cluster,
 		}, e2eCtx.E2EConfig.GetIntervals("", "wait-delete-cluster")...)
 
-		By("Deleting AWS static credentials")
+		ginkgo.By("Deleting AWS static credentials")
 		shared.CleanupStaticCredentials(ctx, namespace, e2eCtx)
 	})
 })
