+++ test/e2e/suites/managed/machine_deployment.go
@@ -22,7 +22,6 @@ import (
 	"context"
 
 	"github.com/aws/aws-sdk-go/aws/client"
-	//. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	corev1 "k8s.io/api/core/v1"
