+++ test/e2e/suites/unmanaged/unmanaged_CAPI_quick_test.go
@@ -20,9 +20,8 @@ package unmanaged
 
 import (
 	"context"
-	"time"
 
-	. "github.com/onsi/ginkgo"
+	"github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	"github.com/gofrs/flock"
@@ -32,31 +31,23 @@ import (
 	capi_e2e "sigs.k8s.io/cluster-api/test/e2e"
 )
 
-
-var _ = Describe("Cluster API E2E tests - unmanaged [PR-Blocking]", func() {
+var _ = ginkgo.Context("[unmanaged] [Cluster API Framework] [smoke] [PR-Blocking]", func() {
 	var (
 		namespace *corev1.Namespace
 		ctx       context.Context
-		specName  = "capi-tests-pr-blocking"
 	)
 
-	BeforeEach(func() {
+	ginkgo.BeforeEach(func() {
 		Expect(e2eCtx.Environment.BootstrapClusterProxy).ToNot(BeNil(), "Invalid argument. BootstrapClusterProxy can't be nil")
 		ctx = context.TODO()
 		// Setup a Namespace where to host objects for this spec and create a watcher for the namespace events.
-		namespace = shared.SetupSpecNamespace(ctx, specName, e2eCtx)
-		Expect(e2eCtx.E2EConfig).ToNot(BeNil(), "Invalid argument. e2eConfig can't be nil when calling %s spec", specName)
-		Expect(e2eCtx.E2EConfig.Variables).To(HaveKey(shared.KubernetesVersion))
-		shared.CreateAWSClusterControllerIdentity(e2eCtx.Environment.BootstrapClusterProxy.GetClient())
-		time.Sleep(5 * time.Second)
+		namespace = shared.SetupSpecNamespace(ctx, "capi-quick-start", e2eCtx)
 	})
 
-	SetDefaultEventuallyTimeout(20 * time.Minute)
-	SetDefaultEventuallyPollingInterval(10 * time.Second)
-	Context("Running the quick-start spec", func() {
+	ginkgo.Describe("Running the quick-start spec", func() {
 		// As the resources cannot be defined by the It() clause in CAPI tests, using the largest values required for all It() tests in this CAPI test.
-		requiredResources := &shared.TestResource{EC2: 2, IGW: 1, NGW: 3, VPC: 1, ClassicLB: 1, EIP: 3}
-		BeforeEach(func() {
+		requiredResources := &shared.TestResource{EC2: 2, IGW: 1, NGW: 1, VPC: 1, ClassicLB: 1, EIP: 3}
+		ginkgo.BeforeEach(func() {
 			requiredResources.WriteRequestedResources(e2eCtx, "capi-quick-start-test")
 			Expect(shared.AcquireResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))).To(Succeed())
 		})
@@ -69,11 +60,11 @@ var _ = Describe("Cluster API E2E tests - unmanaged [PR-Blocking]", func() {
 				SkipCleanup:           e2eCtx.Settings.SkipCleanup,
 			}
 		})
-		AfterEach(func() {
+		ginkgo.AfterEach(func() {
 			shared.ReleaseResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))
 		})
 	})
-	AfterEach(func() {
+	ginkgo.AfterEach(func() {
 		// Dumps all the resources in the spec namespace, then cleanups the cluster object and the spec namespace itself.
 		shared.DumpSpecResourcesAndCleanup(ctx, "", namespace, e2eCtx)
 	})
