+++ test/e2e/suites/unmanaged/unmanaged_CAPI_test.go
@@ -20,44 +20,30 @@ package unmanaged
 
 import (
 	"context"
-	"time"
 
-	. "github.com/onsi/ginkgo"
+	"github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	"github.com/gofrs/flock"
 	"github.com/onsi/ginkgo/config"
-	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/cluster-api-provider-aws/test/e2e/shared"
 	capi_e2e "sigs.k8s.io/cluster-api/test/e2e"
 	"sigs.k8s.io/cluster-api/test/framework/clusterctl"
 )
 
-var _ = Describe("Cluster API E2E tests - unmanaged", func() {
-	var (
-		namespace *corev1.Namespace
-		ctx       context.Context
-		specName  = "capi-tests"
-	)
+var _ = ginkgo.Context("[unmanaged] [Cluster API Framework]", func() {
 
-	BeforeEach(func() {
+	ginkgo.BeforeEach(func() {
 		Expect(e2eCtx.Environment.BootstrapClusterProxy).ToNot(BeNil(), "Invalid argument. BootstrapClusterProxy can't be nil")
-		ctx = context.TODO()
 		// Setup a Namespace where to host objects for this spec and create a watcher for the namespace events.
-		namespace = shared.SetupSpecNamespace(ctx, specName, e2eCtx)
-		Expect(e2eCtx.E2EConfig).ToNot(BeNil(), "Invalid argument. e2eConfig can't be nil when calling %s spec", specName)
-		Expect(e2eCtx.E2EConfig.Variables).To(HaveKey(shared.KubernetesVersion))
-		shared.CreateAWSClusterControllerIdentity(e2eCtx.Environment.BootstrapClusterProxy.GetClient())
-		time.Sleep(5 * time.Second)
 	})
 
-	SetDefaultEventuallyTimeout(20 * time.Minute)
-	SetDefaultEventuallyPollingInterval(10 * time.Second)
-	Context("Running the KCP upgrade spec - a single control plane cluster", func() {
+	// DEPRECATED. Should be replaced with the conformance upgrade spec
+	ginkgo.Describe("KCP Upgrade Spec - Single Control Plane Cluster", func() {
 		// As the resources cannot be defined by the It() clause in CAPI tests, using the largest values required for all It() tests in this CAPI test.
-		requiredResources := &shared.TestResource{EC2: 2, IGW: 1, NGW: 3, VPC: 1, ClassicLB: 1, EIP: 3}
-		BeforeEach(func() {
-			requiredResources.WriteRequestedResources(e2eCtx, "capi-kcp-single-upgrade-test")
+		requiredResources := &shared.TestResource{EC2: 2, IGW: 1, NGW: 1, VPC: 1, ClassicLB: 1, EIP: 3}
+		ginkgo.BeforeEach(func() {
+			requiredResources.WriteRequestedResources(e2eCtx, "capi-kcp-single-cp-upgrade-test")
 			Expect(shared.AcquireResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))).To(Succeed())
 		})
 		capi_e2e.KCPUpgradeSpec(context.TODO(), func() capi_e2e.KCPUpgradeSpecInput {
@@ -72,33 +58,12 @@ var _ = Describe("Cluster API E2E tests - unmanaged", func() {
 			}
 		})
 	})
-	// TODO: Enable when there are more resources in the test infra
-	//Context("Running the KCP upgrade spec - HA cluster", func() {
-	//	// As the resources cannot be defined by the It() clause in CAPI tests, using the largest values required for all It() tests in this CAPI test.
-	//	requiredResources := &shared.TestResource{EC2: 4, IGW: 1, NGW: 3, VPC: 1, ClassicLB: 1, EIP: 3}
-	//	BeforeEach(func() {
-	//		requiredResources.WriteRequestedResources(e2eCtx, "capi-kcp-ha-upgrade-test")
-	//		Expect(shared.AcquireResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))).To(Succeed())
-	//	})
-	//	capi_e2e.KCPUpgradeSpec(context.TODO(), func() capi_e2e.KCPUpgradeSpecInput {
-	//		return capi_e2e.KCPUpgradeSpecInput{
-	//			E2EConfig:                e2eCtx.E2EConfig,
-	//			ClusterctlConfigPath:     e2eCtx.Environment.ClusterctlConfigPath,
-	//			BootstrapClusterProxy:    e2eCtx.Environment.BootstrapClusterProxy,
-	//			Flavor:                   clusterctl.DefaultFlavor,
-	//			ControlPlaneMachineCount: 3,
-	//			ArtifactFolder:           e2eCtx.Settings.ArtifactFolder,
-	//			SkipCleanup:              e2eCtx.Settings.SkipCleanup,
-	//		}
-	//	})
-	//	AfterEach(func() {
-	//		shared.ReleaseResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))
-	//	})
-	//})
-	Context("Running the KCP upgrade spec - HA cluster using scale in rollout", func() {
+
+	// DEPRECATED. Should be replaced with the conformance upgrade spec
+	ginkgo.Describe("KCP Upgrade Spec - HA Control Plane Cluster using Scale-In", func() {
 		// As the resources cannot be defined by the It() clause in CAPI tests, using the largest values required for all It() tests in this CAPI test.
-		requiredResources := &shared.TestResource{EC2: 4, IGW: 1, NGW: 1, VPC: 1, ClassicLB: 1, EIP: 1}
-		BeforeEach(func() {
+		requiredResources := &shared.TestResource{EC2: 4, IGW: 1, NGW: 1, VPC: 1, ClassicLB: 1, EIP: 3}
+		ginkgo.BeforeEach(func() {
 			requiredResources.WriteRequestedResources(e2eCtx, "capi-kcp-scale-in-upgrade-test")
 			Expect(shared.AcquireResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))).To(Succeed())
 		})
@@ -113,35 +78,15 @@ var _ = Describe("Cluster API E2E tests - unmanaged", func() {
 				SkipCleanup:              e2eCtx.Settings.SkipCleanup,
 			}
 		})
-		AfterEach(func() {
+		ginkgo.AfterEach(func() {
 			shared.ReleaseResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))
 		})
 	})
-	Context("Running the MachineDeployment upgrade spec", func() {
-		// As the resources cannot be defined by the It() clause in CAPI tests, using the largest values required for all It() tests in this CAPI test.
-		requiredResources := &shared.TestResource{EC2: 2, IGW: 1, NGW: 3, VPC: 1, ClassicLB: 1, EIP: 3}
-		BeforeEach(func() {
-			requiredResources.WriteRequestedResources(e2eCtx, "capi-machinedeployment-upgrade-test")
-			Expect(shared.AcquireResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))).To(Succeed())
-		})
 
-		capi_e2e.MachineDeploymentUpgradesSpec(context.TODO(), func() capi_e2e.MachineDeploymentUpgradesSpecInput {
-			return capi_e2e.MachineDeploymentUpgradesSpecInput{
-				E2EConfig:             e2eCtx.E2EConfig,
-				ClusterctlConfigPath:  e2eCtx.Environment.ClusterctlConfigPath,
-				BootstrapClusterProxy: e2eCtx.Environment.BootstrapClusterProxy,
-				ArtifactFolder:        e2eCtx.Settings.ArtifactFolder,
-				SkipCleanup:           e2eCtx.Settings.SkipCleanup,
-			}
-		})
-		AfterEach(func() {
-			shared.ReleaseResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))
-		})
-	})
-	Context("Running the MachineRemediation spec", func() {
+	ginkgo.Describe("Machine Remediation Spec", func() {
 		// As the resources cannot be defined by the It() clause in CAPI tests, using the largest values required for all It() tests in this CAPI test.
 		requiredResources := &shared.TestResource{EC2: 4, IGW: 1, NGW: 3, VPC: 1, ClassicLB: 1, EIP: 3}
-		BeforeEach(func() {
+		ginkgo.BeforeEach(func() {
 			requiredResources.WriteRequestedResources(e2eCtx, "capi-remediation-test")
 			Expect(shared.AcquireResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))).To(Succeed())
 		})
@@ -155,14 +100,15 @@ var _ = Describe("Cluster API E2E tests - unmanaged", func() {
 				SkipCleanup:           e2eCtx.Settings.SkipCleanup,
 			}
 		})
-		AfterEach(func() {
+		ginkgo.AfterEach(func() {
 			shared.ReleaseResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))
 		})
 	})
-	Context("Running the Machine pool spec", func() {
+
+	ginkgo.Describe("Machine Pool Spec", func() {
 		// As the resources cannot be defined by the It() clause in CAPI tests, using the largest values required for all It() tests in this CAPI test.
-		requiredResources := &shared.TestResource{EC2: 4, IGW: 1, NGW: 1, VPC: 1, ClassicLB: 1, EIP: 1}
-		BeforeEach(func() {
+		requiredResources := &shared.TestResource{EC2: 4, IGW: 1, NGW: 1, VPC: 1, ClassicLB: 1, EIP: 3}
+		ginkgo.BeforeEach(func() {
 			requiredResources.WriteRequestedResources(e2eCtx, "capi-machinepool-test")
 			Expect(shared.AcquireResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))).To(Succeed())
 		})
@@ -176,13 +122,8 @@ var _ = Describe("Cluster API E2E tests - unmanaged", func() {
 				SkipCleanup:           e2eCtx.Settings.SkipCleanup,
 			}
 		})
-		AfterEach(func() {
+		ginkgo.AfterEach(func() {
 			shared.ReleaseResources(requiredResources, config.GinkgoConfig.ParallelNode, flock.New(shared.ResourceQuotaFilePath))
 		})
-	})
-
-	AfterEach(func() {
-		// Dumps all the resources in the spec namespace, then cleanups the cluster object and the spec namespace itself.
-		shared.DumpSpecResourcesAndCleanup(ctx, "", namespace, e2eCtx)
 	})
 })
