+++ versions.mk
@@ -16,3 +16,5 @@ MDBOOK_VERSION := v0.4.5
 GOLANG_VERSION := 1.16.2
 PLANTUML_VERSION := 1.2020.16
 GH_VERSION := 1.2.0
+CERT_MANAGER_VERSION := v1.1.0
+CAPI_VERSION := v0.4.0
